Factor out {Session -> IsolatedEnvironment}
